Add files via upload
This is a pre-deployment version of the GenerateEmailNotificationLists class. It currently contains code to iterate through an array of email addresses, sending emails individually (to maintain individual privacy while still having a record of each recipient's address) via SMTP and using the PHP imap extension in order to save sent messages to a webmail account.